Renamed wrong namespace. Removed unused using directives #93
